Add files via upload
This course is a good place to start if you want to become a web developer. You will learn about the day-to-day responsibilities of a web developer and get a general understanding of the core and underlying technologies that power the internet. You will learn how front-end developers create websites and applications that work well and are easy to maintain. 

You’ll be introduced to the core web development technologies like HTML and CSS and get opportunities to practice using them. You will also be introduced to modern UI frameworks such as Bootstrap and React that make it easy to create interactive user experiences. 

By the end of the course, you will be able to: 
- Describe the front-end developer role 
- Explain the core and underlying technologies that power the internet 
- Use HTML to create a simple webpage 
- Use CSS to control the appearance of a simple webpage 
- Explain what React is 
- Describe the applications and characteristics of the most popular UI frameworks .