Add files via upload
React is a powerful JavaScript library that you can use to build user interfaces for web and mobile applications (apps). In this course, you will explore the fundamental concepts that underpin the React library and learn the basic skills required to build a simple, fast, and scalable app.

By the end of this course, you will be able to:
•	Use reusable components to render views where data changes over time
•	Create more scalable and maintainable websites and apps 
•	Use props to pass data between components 
•	Create dynamic and interactive web pages and apps
•	Use forms to allow users to interact with the web page 
•	Build an application in React

You’ll gain experience with the following tools and software: 
•	React.js
•	JSX
•	React
•	HTML, CSS and JavaScript 
•	VSCode

You will be able to leverage the potential of this course to develop new skills, improve productivity, act effectively with data and boost your career.

This is a beginner course for learners who would like to prepare themselves for a career in mobile development. To succeed in this course, you do not need prior development experience, only basic internet navigation skills and an eagerness to get started with coding.